requirements.txt
Django>=3.0.3,<3.0.4
djangorestframework>=3.11.0,<3.12.0
django-reset-migrations==0.4.0 
psycopg2>=2.8.5,<2.9.0
Pillow>=7.1.1,<7.1.2

flake8>=3.7.9,<3.8.0